sql-agent-project/requirements.txt
# Core API framework
fastapi
uvicorn

# Database
psycopg2-binary

# AI / LangGraph
langgraph
langchain-openai
langchain-community

# Configuration
python-dotenv

# UI
streamlit
requests
pandas